requirements_lock.txt
byml==2.4.4
libyaz0==0.5
pip==22.0.2
PyYAML==6.0.1
SarcLib==0.3
cgal==5.6.post202309051110
## The following requirements were added by pip freeze:
alembic==1.12.0
amqp==5.1.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
command-not-found==0.3
cryptography==41.0.3
dbus-python==1.2.18
debtcollector==2.5.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.2
docopt==0.6.2
dogpile.cache==1.2.2
elementpath==4.1.5
eventlet==0.33.3
fasteners==0.18
fixtures==4.1.0
Flask==2.3.3
Flask-RESTful==0.3.10
futurist==2.4.1
greenlet==2.0.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iso8601==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.5.0
keystone==23.0.0
keystone-engine==0.9.2
keystoneauth1==5.3.0
keystonemiddleware==10.4.1
kombu==5.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==8.10.0
msgpack==1.0.5
netaddr==0.8.0
netifaces==0.11.0
numpy==1.25.2
oauthlib==3.2.0
oead==1.2.7
os-service-types==1.7.0
oslo.cache==3.5.0
oslo.concurrency==5.2.0
oslo.config==9.2.0
oslo.context==5.2.0
oslo.db==14.0.0
oslo.i18n==6.1.0
oslo.log==5.3.0
oslo.messaging==14.4.0
oslo.metrics==0.7.0
oslo.middleware==5.2.0
oslo.policy==4.2.1
oslo.serialization==5.2.0
oslo.service==3.2.0
oslo.upgradecheck==2.2.0
oslo.utils==6.2.1
osprofiler==4.1.0
packaging==23.1
passlib==1.7.4
Paste==3.5.3
PasteDeploy==3.0.1
pbr==5.11.1
pipreqs==0.4.13
prettytable==3.8.0
prometheus-client==0.17.1
pycadf==3.1.1
pycparser==2.21
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pysaml2==7.4.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-keystoneclient==5.1.0
pytz==2023.3
referencing==0.30.2
repoze.lru==0.7
requests==2.31.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.10.2
scrypt==0.8.20
SecretStorage==3.3.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
sqlalchemy-migrate==0.13.0
sqlparse==0.4.4
statsd==4.0.1
stevedore==5.1.0
systemd-python==234
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.6.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
WebOb==1.8.7
Werkzeug==2.3.7
wrapt==1.15.0
xmlschema==2.4.0
yappi==1.4.0
yarg==0.1.9
zipp==1.0.0
